// packages/sdk/api/workflow/builders/node-builder.ts
export class NodeBuilder extends BaseBuilder<StaticNode> {
// ...
  build(): StaticNode {
// ...
    return {
      id: this._id,
      type: this._type,
      name: this._name || this._id,
      config: this._config as StaticNode["config"],
    } as StaticNode;
  }
// ...
}